add yield
Don't forget the yield in our application view! This piece of code
allows the content of the other views to show up but without having the
trouble of managing many different pages with the same stuff (like
menus and stylesheet loading, etc.)! If that's confusing right now,
don't sweat, well dig deeper into it later on.