fix migration methods
Database migrations can be tricky, but let's read this through for each
migration.

For students, we want to create a table called "students", in this
table we want an attribute column (the attribute will be a string)
called "name", we then want a reference column to a house id (Remember
the belongs_to relationship? This id here will correspond to the
appropriate house the student belongs to in the houses table.) and,
finally, we add timestamps of when our students are created or updated
in the table (which is never a bad idea to have).

For houses, we create a table called "houses" and add an string
attribute column of the house's name called "house_name". You could of
named it "name" as well, like the students table above, but it's more
clear to call them name and house_name respectively, especially while
you're starting out in rails.