generate schema.rb file
This schema.rb file was created when we ran the database migrations --
rake db:migrate -- command. On some projects you won't check this file
into your version controll (git) but it's on a project by project
basis. Likewise, we ran rake db:seed to seed the the houses table with
our four houses of Hogwarts.