edit houses controller show action and houses show view
In the show action in the controller, we want to make available to the
view an instance variable containing the attributes of the particular
house we are trying to show.

In the show view, make sure you're calling the correct attributes
"house_name" (a house attribute) and "name" (a student attribute) for
the correct variables.