edit students controller for creating a new student
Let's get into the nitty gritty of the students controller.

The create action in the controller doesn't have it's own view.
Instead, it's job is to receive the data entered from the form in the
students new view and send it to the Student model for saving. It then
redirects the page to the students index view via a named route
"students_path".

In the student show action, the @student instance variable allows us to
display to the students show view attributes like their name and which
house they belong to.

The @student instance variable in the new action is used in the form in
the students new view.

I cleaned up a bit of your formatting in the students controller. It's
important to keep clean formatting, not only to keep yourself organized
and efficient but also for anyone else who might look at or work with
your code.