add sorting hat feature and ease navigation with more links
Now we have a working Sorting Hat for when a new student gets created!
We also have basic validation to make sure we're not creating duplicate
students (check out the line of code that begins with "validates" in
the Student model). I've add a flash notice system to let the user
adding students know what the Sorting Hat's recomendation was or to
inform the user that their trying to create a dublicate student.
Lastly, I've added a couple more links in the views to help with
navigating through the site. Think about how we might go about adding a
menu to every page to make navigation even easier.

I leave it as an exercise to you to see how I created the sorting hat
feature (Hint: It's only two lines of code!) and try to come up with
your own way to do it! There's very often way more than one way to do
things here in Rails!